Development

+++ gerobug_dashboard/dashboards/views.py
@@ -99,11 +99,17 @@ class ReportDelete(LoginRequiredMixin,DeleteView):
     def delete(self, *args, **kwargs):
         self.object = self.get_object()
         super().delete(*args, **kwargs)
+        #self.object.delete()
 
         # DELETE ALL CHILD UAN OBJECT
-        BugReportUpdate.objects.filter(report_id=self.object.report_id).delete()
-        BugReportAppeal.objects.filter(report_id=self.object.report_id).delete()
-        BugReportNDA.objects.filter(report_id=self.object.report_id).delete()  
+        if BugReportUpdate.objects.filter(report_id=self.object.report_id).exists():
+            BugReportUpdate.objects.filter(report_id=self.object.report_id).delete()
+
+        if BugReportAppeal.objects.filter(report_id=self.object.report_id).exists():
+            BugReportAppeal.objects.filter(report_id=self.object.report_id).delete()
+        
+        if BugReportNDA.objects.filter(report_id=self.object.report_id).exists():
+            BugReportNDA.objects.filter(report_id=self.object.report_id).delete() 
 
     def get_success_url(self):
         if platform == "win32":
@@ -566,7 +572,10 @@ def halloffame(request,):
     return render(request, 'halloffame.html',{'bughunters':bughunters})
 
 def notfound_404(request, exception):
-    return render(request, 'notfound.html', status=404)
+    return render(request, '404.html', status=404)
+
+# def error_500(request, exception):
+#     return render(request, '500.html', status=500)
 
 class Themes(TemplateView):
     template_name = 'theme.css'
